Adressed mr reviews for code organization.
